fix: add server-side access control to admin layout
The admin layout was a client component without access control checks,
allowing anyone to view the admin UI even though the APIs were protected.

Changes:
- Converted admin layout to server component
- Added server-side authentication and admin role verification
- Created separate AdminSidebar client component for navigation
- Non-admin users now see access denied message
- Unauthenticated users are redirected to sign-in

Fixes #10
