+++ src/app/admin/layout.tsx
@@ -1,66 +1,49 @@
 /**
- * Admin Layout
- * Sidebar navigation for admin section
+ * Admin Layout - Server Component
+ * Checks admin access before rendering admin UI
  */
 
-'use client';
+import { redirect } from 'next/navigation';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
+import { UserManagementService } from '@/lib/admin/user-management-service';
+import { AdminSidebar } from './admin-sidebar';
 
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-
-export default function AdminLayout({
+export default async function AdminLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const pathname = usePathname();
+  const session = await getServerSession(authOptions);
 
-  const navItems = [
-    { href: '/admin/users', label: 'Users', icon: '👥' },
-    { href: '/admin/requests', label: 'Access Requests', icon: '📧' },
-  ];
+  // Redirect to sign in if not authenticated
+  if (!session?.user?.email) {
+    redirect('/api/auth/signin');
+  }
 
+  // Check if user is admin
+  const isAdmin = await UserManagementService.isAdmin(session.user.email);
+
+  // Show access denied if not admin
+  if (!isAdmin) {
+    return (
+      <div className="container mx-auto max-w-4xl px-4 py-12 mt-20">
+        <div className="rounded-xl border border-red-500/50 bg-red-500/10 p-6 text-center">
+          <h1 className="text-2xl font-bold text-red-400 mb-2">Access Denied</h1>
+          <p className="text-red-400/80">Admin role required to access this area.</p>
+          <p className="text-white/60 mt-4 text-sm">
+            If you believe you should have access, please contact the administrator.
+          </p>
+        </div>
+      </div>
+    );
+  }
+
+  // Render admin UI for authorized users
   return (
     <div className="flex min-h-screen">
-      {/* Sidebar */}
-      <aside className="fixed left-0 top-20 bottom-0 w-64 border-r border-white/10 bg-slate-950/95 p-6">
-        <div className="space-y-6">
-          <div>
-            <h2 className="mb-4 text-sm font-semibold uppercase tracking-wider text-white/60">
-              Admin Panel
-            </h2>
-            <nav className="space-y-2">
-              {navItems.map((item) => {
-                const isActive = pathname === item.href;
-                return (
-                  <Link
-                    key={item.href}
-                    href={item.href}
-                    className={`flex items-center gap-3 rounded-lg px-4 py-3 text-sm font-medium transition ${
-                      isActive
-                        ? 'bg-cyan-500/20 text-cyan-400'
-                        : 'text-white/70 hover:bg-white/5 hover:text-white'
-                    }`}
-                  >
-                    <span className="text-lg">{item.icon}</span>
-                    <span>{item.label}</span>
-                  </Link>
-                );
-              })}
-            </nav>
-          </div>
-
-          <div className="border-t border-white/10 pt-6">
-            <Link
-              href="/"
-              className="flex items-center gap-2 text-sm text-white/50 transition hover:text-white"
-            >
-              <span>←</span>
-              <span>Back to Site</span>
-            </Link>
-          </div>
-        </div>
-      </aside>
+      {/* Client-side sidebar for navigation */}
+      <AdminSidebar />
 
       {/* Main Content */}
       <main className="ml-64 flex-1 pt-20">
